+++ software-project-management-ui/src/api-services/issueServices.tsx
@@ -7,43 +7,48 @@ import { handleTokenExpired, RefreshToken } from "./AuthServices";
 // -----------------------------------Issue-----------------------------------
 
 export const fetchIssue = async () => {
-    try {
-        // await RefreshToken();
-        const response = await axios.get("/issue/get-all/673f4fde2856710bc3634f7e", {
-            headers: {
-                Authorization: `Bearer ${getAccessTokenFromCookie()}`,
-            },
-            withCredentials: true,
-        });
-        return response.data;
-    } catch (error) {
-        console.log(error);
-        await handleTokenExpired(error);
-    }
+	try {
+		// await RefreshToken();
+		const response = await axios.get(
+			"/issue/get-all/6740b872a950648ea070aa07",
+			{
+				headers: {
+					Authorization: `Bearer ${getAccessTokenFromCookie()}`,
+				},
+				withCredentials: true,
+			}
+		);
+		return response.data;
+	} catch (error) {
+		console.log(error);
+		await handleTokenExpired(error);
+	}
 };
 
 export const createIssue = async (data: any) => {
-    try {
-        let { summary } = data;
-        let issueType = data?.issueType || "Story";
+	try {
+		let { summary } = data;
+		let issueType = data?.issueType || "Story";
 
-        // await RefreshToken();
-        const response = await axios.post(
-            "/issue/673f4fde2856710bc3634f7e",
-            {
-                issueType,
-                summary,
-            },
-            {
-                headers: {
-                    Authorization: `Bearer ${getAccessTokenFromCookie()}`,
-                },
-                withCredentials: true,
-            }
-        );
-        return response.data;
-    } catch (error) {
-        console.log(error);
-        await handleTokenExpired(error);
-    }
+		// await RefreshToken();
+		const response = await axios.post(
+			"/issue/6740b872a950648ea070aa07",
+			{
+				issueType,
+				summary,
+			},
+			{
+				headers: {
+					Authorization: `Bearer ${getAccessTokenFromCookie()}`,
+				},
+				withCredentials: true,
+			}
+		);
+		console.log(issueType);
+		console.log(summary);
+		return response.data;
+	} catch (error) {
+		console.log(error);
+		await handleTokenExpired(error);
+	}
 };
