+++ software-project-management-ui/src/api-services/sprintService.tsx
@@ -7,53 +7,60 @@ import { handleTokenExpired, RefreshToken } from "./AuthServices";
 // -----------------------------------Issue-----------------------------------
 
 export const fetchAllSprint = async () => {
-    try {
-        // await RefreshToken();
-        const response = await axios.get("/sprint/get-all/673f4fde2856710bc3634f7e", {
-            headers: {
-                Authorization: `Bearer ${getAccessTokenFromCookie()}`,
-            },
-            withCredentials: true,
-        });
-        return response.data;
-    } catch (error) {
-        console.log(error);
-        await handleTokenExpired(error);
-    }
+	try {
+		// await RefreshToken();
+		const response = await axios.get(
+			"/sprint/get-all/6740b872a950648ea070aa07",
+			{
+				headers: {
+					Authorization: `Bearer ${getAccessTokenFromCookie()}`,
+				},
+				withCredentials: true,
+			}
+		);
+		return response.data;
+	} catch (error) {
+		console.log(error);
+		await handleTokenExpired(error);
+	}
 };
 
 export const createSprint = async () => {
-    try {
-        // await RefreshToken();
-        const response = await axios.post(
-            "/sprint/673f4fde2856710bc3634f7e",
-            {},
-            {
-                headers: {
-                    Authorization: `Bearer ${getAccessTokenFromCookie()}`,
-                },
-                withCredentials: true,
-            }
-        );
-        return response.data;
-    } catch (error) {
-        console.log(error);
-        await handleTokenExpired(error);
-    }
+	try {
+		// await RefreshToken();
+		const response = await axios.post(
+			"/sprint/6740b872a950648ea070aa07",
+			{},
+			{
+				headers: {
+					Authorization: `Bearer ${getAccessTokenFromCookie()}`,
+				},
+				withCredentials: true,
+			}
+		);
+		return response.data;
+	} catch (error) {
+		console.log(error);
+		await handleTokenExpired(error);
+	}
 };
 
 export const deleteSprint = async (sprintId: string) => {
-    try {
-        // await RefreshToken();
-        const response = await axios.delete(`/sprint/673f4fde2856710bc3634f7e/${sprintId}`, {
-            headers: {
-                Authorization: `Bearer ${getAccessTokenFromCookie()}`,
-            },
-            withCredentials: true,
-        });
-        return response.data;
-    } catch (error) {
-        console.log(error);
-        await handleTokenExpired(error);
-    }
+	try {
+		// await RefreshToken();
+		const response = await axios.delete(
+			`/sprint/6740b872a950648ea070aa07/${sprintId}`,
+			{
+				headers: {
+					Authorization: `Bearer ${getAccessTokenFromCookie()}`,
+				},
+				withCredentials: true,
+			}
+		);
+		console.log(sprintId);
+		return response.data;
+	} catch (error) {
+		console.log(error);
+		await handleTokenExpired(error);
+	}
 };
