+++ software-project-management-ui/src/app/sine/backlog/page.tsx
@@ -3,26 +3,24 @@ import * as React from "react";
 import Grid from "@mui/material/Grid";
 import { deepOrange, deepPurple } from "@mui/material/colors";
 import {
-    AvatarGroup,
-    Breadcrumbs,
-    Button,
-    Dialog,
-    Fade,
-    IconButton,
-    Input,
-    LinearProgress,
-    Link,
-    Select,
-    Menu,
-    MenuItem,
-    Paper,
-    Table,
-    TableBody,
-    TableCell,
-    TableRow,
-    Tooltip,
-    FormControlLabel,
-    Switch,
+	AvatarGroup,
+	Breadcrumbs,
+	Button,
+	Dialog,
+	Fade,
+	IconButton,
+	Link,
+	Select,
+	Menu,
+	MenuItem,
+	Table,
+	TableBody,
+	TableCell,
+	TableRow,
+	Tooltip,
+	FormControlLabel,
+	Switch,
+	LinearProgress,
 } from "@mui/material";
 import styles from "@/components/Apps/FileManager/Sidebar/SearchForm/Search.module.css";
 import { Card, Typography, Avatar, Badge, styled, Box } from "@mui/material";
@@ -50,15 +48,13 @@ import {
 import { sortableKeyboardCoordinates } from "@dnd-kit/sortable";
 
 import { arrayMove } from "@dnd-kit/sortable";
-import ExampleDND from "../drag&drop/page";
 import "../drag&drop/component/Column/Column.css";
 import { restrictToVerticalAxis } from "@dnd-kit/modifiers";
-import { Backlog } from "./BacklogList/BackLog/Backlog";
-import { toast } from "react-toastify";
 
 import * as issueService from "@/api-services/issueServices";
 import * as sprintService from "@/api-services/sprintService";
-import { set } from "react-hook-form";
+import { CreateBacklog } from "./CreateBacklog/CreateBacklog";
+import { toast } from "react-toastify";
 
 type Issue = {
 	_id: string;
@@ -67,11 +63,16 @@ type Issue = {
 	// Add other properties if necessary
 };
 
+type Sprint = {
+	name: string; // Định nghĩa kiểu cho Sprint
+	_id: string; // Thêm thuộc tính _id
+	// Thêm các thuộc tính khác nếu cần
+};
+
 export default function Page({ projectName }: { projectName: string }) {
 	const [issue, setIssue] = useState<Issue[]>([]);
 	const [update, setUpdate] = useState(false);
-	const [fetchedSprint, setFetchedSprint] = useState([]);
-	const [issueName, setIssueName] = useState("");
+	const [fetchedSprint, setFetchedSprint] = useState<Sprint[]>([]);
 	useEffect(() => {
 		const fetchAPI = async () => {
 			const result = await sprintService.fetchAllSprint();
@@ -81,8 +82,8 @@ export default function Page({ projectName }: { projectName: string }) {
 	}, [update]);
 
 	useEffect(() => {
-		const sprintNames = fetchedSprint.map((sprint) => sprint?.name);
-		setSprints(sprintNames);
+		const sprintData = Array.isArray(fetchedSprint) ? fetchedSprint : [];
+		setSprints(sprintData);
 	}, [fetchedSprint]);
 
 	useEffect(() => {
@@ -94,17 +95,23 @@ export default function Page({ projectName }: { projectName: string }) {
 	}, [update]);
 
 	useEffect(() => {
-		const mappedBacklogs = issue.map((item) => ({
-			id: item.key, // Use the same ID
-			title: item.key, // Map summary to title
-			description: item.summary, // Example mapping for description
-		}));
+		const mappedBacklogs = Array.isArray(issue)
+			? issue.map((item) => ({
+					id: item.key, // Use the same ID
+					title: item.key, // Map summary to title
+					description: item.summary, // Example mapping for description
+				}))
+			: [];
 		setBacklogs(mappedBacklogs);
 	}, [issue]);
 
-	const handleDeleteSprint = async () => {
-		await sprintService.deleteSprint("1");
-		setUpdate(!update);
+	const handleDeleteSprint = async (sprintId: string) => {
+		if (sprintId) {
+			setOpenNotification(false);
+			await sprintService.deleteSprint(sprintId);
+			toast.success("Delete Sprint Successful!");
+			setUpdate(!update);
+		}
 	};
 
 	const breadcrumbs = [
@@ -159,8 +166,7 @@ export default function Page({ projectName }: { projectName: string }) {
 
 	const [isEpicVisible, setIsEpicVisible] = React.useState(true);
 	const [expanded, setExpanded] = React.useState<string | string[]>([]);
-	const [sprints, setSprints] = React.useState<string[]>([]);
-	const [newbacklogs, setNewBacklogs] = React.useState<string[]>([]);
+	const [sprints, setSprints] = useState<Sprint[]>([]);
 
 	const handleAccordionChange =
 		(panel: string) => (event: React.SyntheticEvent) => {
@@ -174,40 +180,21 @@ export default function Page({ projectName }: { projectName: string }) {
 						: [panel]
 			);
 		};
+	const [loading, setLoading] = useState(false);
 
 	const handleCreateSprint = async () => {
-		let sprint = await sprintService.createSprint();
-		console.log(sprint);
-		setUpdate(!update);
+		setLoading(true);
+		await sprintService.createSprint();
+		// console.log(sprint);
+		setTimeout(() => {
+			setUpdate(!update);
+			toast.success("Create Sprint Successful!");
+			setLoading(false);
+		}, 1000);
+		console.log(sprints);
 		// setSprints((prev) => [...prev, `Sprint ${prev.length + 1}`]);
 	};
 
-	const handleCreateBacklog = () => {
-		setNewBacklogs((prev) => [...prev, `Backlog ${prev.length + 1}`]);
-	};
-
-	const [loading, setLoading] = useState(false);
-
-	const handleBacklogSubmit = async () => {
-		setLoading(true);
-		let issue = await issueService.createIssue({ summary: issueName });
-		setTimeout(() => {
-			toast.success("Create Backlog Successful!");
-			setLoading(false);
-			setIssueName(""); // Clear the input after submission
-		}, 2000);
-	};
-
-	const Item = styled(Paper)(({ theme }) => ({
-		backgroundColor: "#fff",
-		...theme.typography.body2,
-		padding: theme.spacing(1),
-		textAlign: "center",
-		color: theme.palette.text.secondary,
-		...theme.applyStyles("dark", {
-			backgroundColor: "#1A2027",
-		}),
-	}));
 	React.useEffect(() => {
 		const handleKeyPress = (event: KeyboardEvent) => {
 			if (event.key === "`") {
@@ -636,397 +623,368 @@ export default function Page({ projectName }: { projectName: string }) {
 						xl={isEpicVisible ? 9 : 12}
 					>
 						{/* <Card
-                            sx={{
-                                boxShadow: "none",
-                                borderRadius: "7px",
-                                mb: "10px",
-                                padding: { xs: "8px", sm: "10px", lg: "15px" },
-                            }}
-                            className="rmui-card"
-                        >
-                            <Box display="flex" alignItems="center" justifyContent="space-between">
-                                <Box display="flex">
-                                    <StartSprintDialog />
-                                </Box>
-                                <Box
-                                    display="flex"
-                                    sx={{
-                                        display: "flex",
-                                        flexDirection: "row",
-                                    }}
-                                >
-                                    <Button
-                                        id="fade-button"
-                                        aria-controls={open ? "fade-menu" : undefined}
-                                        aria-haspopup="true"
-                                        aria-expanded={open ? "true" : undefined}
-                                        onClick={handleClick}
-                                    >
-                                        <span className="material-symbols-outlined">
-                                            more_horiz
-                                        </span>
-                                    </Button>
-                                    <Menu
-                                        id="fade-menu"
-                                        MenuListProps={{
-                                            "aria-labelledby": "fade-button",
-                                        }}
-                                        anchorEl={anchorEl}
-                                        open={open}
-                                        onClose={handleClose}
-                                        TransitionComponent={Fade}
-                                    >
-                                        <MenuItem onClick={handleClose}>Project settings</MenuItem>
-                                        <MenuItem onClick={handleClickOpenNotification}>
-                                            Move to trash
-                                        </MenuItem>
-                                    </Menu>
-                                </Box>
-                            </Box>
+							sx={{
+								boxShadow: "none",
+								borderRadius: "7px",
+								mb: "10px",
+								padding: { xs: "8px", sm: "10px", lg: "15px" },
+							}}
+							className="rmui-card"
+						>
+							<Box
+								display="flex"
+								alignItems="center"
+								justifyContent="space-between"
+							>
+								<Box display="flex">
+									<StartSprintDialog />
+								</Box>
+								<Box
+									display="flex"
+									sx={{
+										display: "flex",
+										flexDirection: "row",
+									}}
+								>
+									<Button
+										id="fade-button"
+										aria-controls={open ? "fade-menu" : undefined}
+										aria-haspopup="true"
+										aria-expanded={open ? "true" : undefined}
+										onClick={handleClick}
+									>
+										<span className="material-symbols-outlined">
+											more_horiz
+										</span>
+									</Button>
+									<Menu
+										id="fade-menu"
+										MenuListProps={{
+											"aria-labelledby": "fade-button",
+										}}
+										anchorEl={anchorEl}
+										open={open}
+										onClose={handleClose}
+										TransitionComponent={Fade}
+									>
+										<MenuItem onClick={handleClose}>Project settings</MenuItem>
+										<MenuItem onClick={handleClickOpenNotification}>
+											Move to trash
+										</MenuItem>
+									</Menu>
+								</Box>
+							</Box>
 
-                            { <Accordion
-                                expanded={
-                                    Array.isArray(expanded)
-                                        ? expanded.includes("panel4")
-                                        : expanded === "panel4"
-                                }
-                                onChange={handleAccordionChange("panel4")}
-                                className="accordionItem"
-                                sx={{
-                                    backgroundColor:
-                                        Array.isArray(expanded) && expanded.includes("panel4")
-                                            ? "#e9ebee"
-                                            : "inherit",
-                                    "&:hover": {
-                                        backgroundColor: "#e9ebee",
-                                    },
-                                    boxShadow: "none",
-                                    border: "none",
-                                    padding: { xs: "0px", sm: "0px", lg: "0px" },
-                                    flexGrow: 1, // Đảm bảo accordion chiếm không gian còn lại
-                                }}
-                            >
-                                <AccordionSummary
-                                    expandIcon={<ExpandMoreIcon />}
-                                    aria-controls="panel4-content"
-                                    id="panel4-header"
-                                    sx={{ fontWeight: "500", fontSize: "15px" }}
-                                >
-                                    FP Sprint 1
-                                </AccordionSummary> */}
-						{/* <AccordionDetails>
-                                    <Stack spacing={1}>
-                                        <AccordionDetails>
-                                            <DndContext
-                                                modifiers={[restrictToVerticalAxis]}
-                                                sensors={sensors}
-                                                collisionDetection={closestCorners}
-                                                onDragStart={handleOnDragStart}
-                                                onDragEnd={(event) =>
-                                                    handleDragEnd(
-                                                        event as {
-                                                            active: { id: string };
-                                                            over: { id: string };
-                                                        }
-                                                    )
-                                                }
-                                            >
-                                                <BacklogList backlogs={backlogs}></BacklogList>
-                                                <DragOverlay dropAnimation={dropAnimation}>
-                                                    {!activeDragItemData && null}
-                                                    {
-                                                        <div
-                                                            className="backlogItem"
-                                                            style={{ padding: "0px 0px 0px 0px" }}
-                                                        >
-                                                            <Table
-                                                                sx={{
-                                                                    borderBottom: "none !important",
-                                                                }}
-                                                            >
-                                                                <TableBody>
-                                                                    <TableRow>
-                                                                        <TableCell
-                                                                            style={{
-                                                                                border: "none",
-                                                                                alignItems:
-                                                                                    "center",
-                                                                                justifyContent:
-                                                                                    "center",
-                                                                            }}
-                                                                        >
-                                                                            <div
-                                                                                style={{
-                                                                                    paddingTop:
-                                                                                        "5px",
-                                                                                    display: "flex",
-                                                                                    justifyContent:
-                                                                                        "center",
-                                                                                }}
-                                                                            >
-                                                                                <svg
-                                                                                    width="20px"
-                                                                                    height="20px"
-                                                                                    style={{
-                                                                                        marginRight:
-                                                                                            "5px",
-                                                                                    }}
-                                                                                    viewBox="0 0 16 16"
-                                                                                    version="1.1"
-                                                                                    xmlns="http://www.w3.org/2000/svg"
-                                                                                >
-                                                                                    <defs></defs>
-                                                                                    <g
-                                                                                        id="Page-1"
-                                                                                        stroke="none"
-                                                                                        strokeWidth="1"
-                                                                                        fill="none"
-                                                                                        fillRule="evenodd"
-                                                                                    >
-                                                                                        <g id="task">
-                                                                                            <g
-                                                                                                id="Task"
-                                                                                                transform="translate(1.000000, 1.000000)"
-                                                                                            >
-                                                                                                <rect
-                                                                                                    id="Rectangle-36"
-                                                                                                    fill="#4BADE8"
-                                                                                                    x="0"
-                                                                                                    y="0"
-                                                                                                    width="14"
-                                                                                                    height="14"
-                                                                                                    rx="2"
-                                                                                                ></rect>
-                                                                                                <g
-                                                                                                    id="Page-1"
-                                                                                                    transform="translate(4.000000, 4.500000)"
-                                                                                                    stroke="#FFFFFF"
-                                                                                                    strokeWidth="2"
-                                                                                                    strokeLinecap="round"
-                                                                                                >
-                                                                                                    <path
-                                                                                                        d="M2,5 L6,0"
-                                                                                                        id="Stroke-1"
-                                                                                                    ></path>
-                                                                                                    <path
-                                                                                                        d="M2,5 L0,3"
-                                                                                                        id="Stroke-3"
-                                                                                                    ></path>
-                                                                                                </g>
-                                                                                            </g>
-                                                                                        </g>
-                                                                                    </g>
-                                                                                </svg>
-                                                                            </div>
-                                                                        </TableCell>
-                                                                        <TableCell
-                                                                            style={{
-                                                                                border: "none",
-                                                                            }}
-                                                                            sx={{ width: "50%" }}
-                                                                        >
-                                                                            <Link
-                                                                                className="hover-underlined"
-                                                                                color="inherit"
-                                                                                href=""
-                                                                            ></Link>
-                                                                        </TableCell>
-                                                                        <TableCell
-                                                                            style={{
-                                                                                border: "none",
-                                                                                display: "flex",
-                                                                                flexDirection:
-                                                                                    "row",
-                                                                            }}
-                                                                        >
-                                                                            <Select
-                                                                                labelId="product-type-label"
-                                                                                id="product-type"
-                                                                                className="epicSelectBg"
-                                                                                size="small"
-                                                                                style={{
-                                                                                    marginRight:
-                                                                                        "5px",
-                                                                                }}
-                                                                                sx={{
-                                                                                    "& fieldset": {
-                                                                                        maxWidth:
-                                                                                            "120px",
-                                                                                    },
-                                                                                    "& .MuiSelect-select":
-                                                                                        {
-                                                                                            overflow:
-                                                                                                "hidden",
-                                                                                            textOverflow:
-                                                                                                "ellipsis",
-                                                                                            whiteSpace:
-                                                                                                "nowrap",
-                                                                                        },
-                                                                                }}
-                                                                            >
-                                                                                <MenuItem value={0}>
-                                                                                    epic
-                                                                                </MenuItem>
-                                                                                <MenuItem value={1}>
-                                                                                    1
-                                                                                </MenuItem>
-                                                                                <MenuItem value={2}>
-                                                                                    2
-                                                                                </MenuItem>
-                                                                                <MenuItem value={3}>
-                                                                                    3
-                                                                                </MenuItem>
-                                                                                <MenuItem value={4}>
-                                                                                    4
-                                                                                </MenuItem>
-                                                                            </Select>
-                                                                            <Select
-                                                                                labelId="product-type-label"
-                                                                                className="progressSelectBg"
-                                                                                id="product-type"
-                                                                                size="small"
-                                                                                sx={{
-                                                                                    "& fieldset":
-                                                                                        {},
-                                                                                    "& .MuiSelect-select":
-                                                                                        {
-                                                                                            overflow:
-                                                                                                "hidden",
-                                                                                            textOverflow:
-                                                                                                "ellipsis",
-                                                                                            whiteSpace:
-                                                                                                "nowrap",
-                                                                                        },
-                                                                                }}
-                                                                            >
-                                                                                <MenuItem value={0}>
-                                                                                    progress
-                                                                                </MenuItem>
-                                                                                <MenuItem value={1}>
-                                                                                    1
-                                                                                </MenuItem>
-                                                                                <MenuItem value={2}>
-                                                                                    2
-                                                                                </MenuItem>
-                                                                                <MenuItem value={3}>
-                                                                                    3
-                                                                                </MenuItem>
-                                                                                <MenuItem value={4}>
-                                                                                    4
-                                                                                </MenuItem>
-                                                                            </Select>
-                                                                        </TableCell>
-                                                                        <TableCell
-                                                                            style={{
-                                                                                border: "none",
-                                                                            }}
-                                                                        ></TableCell>
-                                                                    </TableRow>
-                                                                </TableBody>
-                                                            </Table>
-                                                        </div>
-                                                    }
-                                                </DragOverlay>
-                                            </DndContext>
-                                        </AccordionDetails>
-                                    </Stack>
-                                </AccordionDetails>
-                            </Accordion> }
-                            <BootstrapDialog
-                                onClose={handleCloseNotification}
-                                aria-labelledby="customized-dialog-title"
-                                open={openNotification}
-                                className="rmu-modal"
-                            >
-                                <Box>
-                                    <Box
-                                        sx={{
-                                            display: "flex",
-                                            justifyContent: "space-between",
-                                            alignItems: "center",
-                                            background: "#ff6666",
-                                            padding: { xs: "15px 20px", md: "25px" },
-                                        }}
-                                        className="rmu-modal-header"
-                                    >
-                                        <Typography
-                                            id="modal-modal-title"
-                                            variant="h6"
-                                            sx={{
-                                                fontWeight: "600",
-                                                fontSize: { xs: "16px", md: "18px" },
-                                                color: "#fff !important",
-                                            }}
-                                            className="text-black"
-                                        >
-                                            Move to Trash
-                                        </Typography>
+							{
+								<Accordion
+									expanded={
+										Array.isArray(expanded)
+											? expanded.includes("panel4")
+											: expanded === "panel4"
+									}
+									onChange={handleAccordionChange("panel4")}
+									className="accordionItem"
+									sx={{
+										backgroundColor:
+											Array.isArray(expanded) && expanded.includes("panel4")
+												? "#e9ebee"
+												: "inherit",
+										"&:hover": {
+											backgroundColor: "#e9ebee",
+										},
+										boxShadow: "none",
+										border: "none",
+										padding: { xs: "0px", sm: "0px", lg: "0px" },
+										flexGrow: 1, // Đảm bảo accordion chiếm không gian còn lại
+									}}
+								>
+									<AccordionSummary
+										expandIcon={<ExpandMoreIcon />}
+										aria-controls="panel4-content"
+										id="panel4-header"
+										sx={{ fontWeight: "500", fontSize: "15px" }}
+									>
+										FP Sprint 1
+									</AccordionSummary>
+									<AccordionDetails>
+										<Stack spacing={1}>
+											<AccordionDetails>
+												<DndContext
+													modifiers={[restrictToVerticalAxis]}
+													sensors={sensors}
+													collisionDetection={closestCorners}
+													onDragStart={handleOnDragStart}
+													onDragEnd={(event) =>
+														handleDragEnd(
+															event as {
+																active: { id: string };
+																over: { id: string };
+															}
+														)
+													}
+												>
+													<BacklogList backlogs={backlogs}></BacklogList>
+													<DragOverlay dropAnimation={dropAnimation}>
+														{!activeDragItemData && null}
+														{
+															<div
+																className="backlogItem"
+																style={{ padding: "0px 0px 0px 0px" }}
+															>
+																<Table
+																	sx={{
+																		borderBottom: "none !important",
+																	}}
+																>
+																	<TableBody>
+																		<TableRow>
+																			<TableCell
+																				style={{
+																					border: "none",
+																					alignItems: "center",
+																					justifyContent: "center",
+																				}}
+																			>
+																				<div
+																					style={{
+																						paddingTop: "5px",
+																						display: "flex",
+																						justifyContent: "center",
+																					}}
+																				>
+																					<svg
+																						width="20px"
+																						height="20px"
+																						style={{
+																							marginRight: "5px",
+																						}}
+																						viewBox="0 0 16 16"
+																						version="1.1"
+																						xmlns="http://www.w3.org/2000/svg"
+																					>
+																						<defs></defs>
+																						<g
+																							id="Page-1"
+																							stroke="none"
+																							strokeWidth="1"
+																							fill="none"
+																							fillRule="evenodd"
+																						>
+																							<g id="task">
+																								<g
+																									id="Task"
+																									transform="translate(1.000000, 1.000000)"
+																								>
+																									<rect
+																										id="Rectangle-36"
+																										fill="#4BADE8"
+																										x="0"
+																										y="0"
+																										width="14"
+																										height="14"
+																										rx="2"
+																									></rect>
+																									<g
+																										id="Page-1"
+																										transform="translate(4.000000, 4.500000)"
+																										stroke="#FFFFFF"
+																										strokeWidth="2"
+																										strokeLinecap="round"
+																									>
+																										<path
+																											d="M2,5 L6,0"
+																											id="Stroke-1"
+																										></path>
+																										<path
+																											d="M2,5 L0,3"
+																											id="Stroke-3"
+																										></path>
+																									</g>
+																								</g>
+																							</g>
+																						</g>
+																					</svg>
+																				</div>
+																			</TableCell>
+																			<TableCell
+																				style={{
+																					border: "none",
+																				}}
+																				sx={{ width: "50%" }}
+																			>
+																				<Link
+																					className="hover-underlined"
+																					color="inherit"
+																					href=""
+																				></Link>
+																			</TableCell>
+																			<TableCell
+																				style={{
+																					border: "none",
+																					display: "flex",
+																					flexDirection: "row",
+																				}}
+																			>
+																				<Select
+																					labelId="product-type-label"
+																					id="product-type"
+																					className="epicSelectBg"
+																					size="small"
+																					style={{
+																						marginRight: "5px",
+																					}}
+																					sx={{
+																						"& fieldset": {
+																							maxWidth: "120px",
+																						},
+																						"& .MuiSelect-select": {
+																							overflow: "hidden",
+																							textOverflow: "ellipsis",
+																							whiteSpace: "nowrap",
+																						},
+																					}}
+																				>
+																					<MenuItem value={0}>epic</MenuItem>
+																					<MenuItem value={1}>1</MenuItem>
+																					<MenuItem value={2}>2</MenuItem>
+																					<MenuItem value={3}>3</MenuItem>
+																					<MenuItem value={4}>4</MenuItem>
+																				</Select>
+																				<Select
+																					labelId="product-type-label"
+																					className="progressSelectBg"
+																					id="product-type"
+																					size="small"
+																					sx={{
+																						"& fieldset": {},
+																						"& .MuiSelect-select": {
+																							overflow: "hidden",
+																							textOverflow: "ellipsis",
+																							whiteSpace: "nowrap",
+																						},
+																					}}
+																				>
+																					<MenuItem value={0}>
+																						progress
+																					</MenuItem>
+																					<MenuItem value={1}>1</MenuItem>
+																					<MenuItem value={2}>2</MenuItem>
+																					<MenuItem value={3}>3</MenuItem>
+																					<MenuItem value={4}>4</MenuItem>
+																				</Select>
+																			</TableCell>
+																			<TableCell
+																				style={{
+																					border: "none",
+																				}}
+																			></TableCell>
+																		</TableRow>
+																	</TableBody>
+																</Table>
+															</div>
+														}
+													</DragOverlay>
+												</DndContext>
+											</AccordionDetails>
+										</Stack>
+									</AccordionDetails>
+								</Accordion>
+							}
+							<BootstrapDialog
+								onClose={handleCloseNotification}
+								aria-labelledby="customized-dialog-title"
+								open={openNotification}
+								className="rmu-modal"
+							>
+								<Box>
+									<Box
+										sx={{
+											display: "flex",
+											justifyContent: "space-between",
+											alignItems: "center",
+											background: "#ff6666",
+											padding: { xs: "15px 20px", md: "25px" },
+										}}
+										className="rmu-modal-header"
+									>
+										<Typography
+											id="modal-modal-title"
+											variant="h6"
+											sx={{
+												fontWeight: "600",
+												fontSize: { xs: "16px", md: "18px" },
+												color: "#fff !important",
+											}}
+											className="text-black"
+										>
+											Move to Trash
+										</Typography>
 
-                                        <IconButton
-                                            aria-label="remove"
-                                            size="small"
-                                            onClick={handleCloseNotification}
-                                        >
-                                            <ClearIcon />
-                                        </IconButton>
-                                    </Box>
+										<IconButton
+											aria-label="remove"
+											size="small"
+											onClick={handleCloseNotification}
+										>
+											<ClearIcon />
+										</IconButton>
+									</Box>
 
-                                    <Box className="rmu-modal-content">
-                                        <Box component="form" noValidate onSubmit={handleSubmit}>
-                                            <Box
-                                                sx={{
-                                                    padding: "25px",
-                                                    borderRadius: "8px",
-                                                }}
-                                                className="bg-white"
-                                            >
-                                                <Grid container alignItems="center" spacing={2}>
-                                                    <Grid item xs={12} mt={1}>
-                                                        <Box
-                                                            sx={{
-                                                                display: "flex",
-                                                                alignItems: "center",
-                                                                gap: "10px",
-                                                            }}
-                                                        >
-                                                            <Button
-                                                                onClick={handleCloseNotification}
-                                                                variant="outlined"
-                                                                color="error"
-                                                                sx={{
-                                                                    textTransform: "capitalize",
-                                                                    borderRadius: "8px",
-                                                                    fontWeight: "500",
-                                                                    fontSize: "13px",
-                                                                    padding: "11px 30px",
-                                                                }}
-                                                            >
-                                                                Cancel
-                                                            </Button>
+									<Box className="rmu-modal-content">
+										<Box component="form" noValidate onSubmit={handleSubmit}>
+											<Box
+												sx={{
+													padding: "25px",
+													borderRadius: "8px",
+												}}
+												className="bg-white"
+											>
+												<Grid container alignItems="center" spacing={2}>
+													<Grid item xs={12} mt={1}>
+														<Box
+															sx={{
+																display: "flex",
+																alignItems: "center",
+																gap: "10px",
+															}}
+														>
+															<Button
+																onClick={handleCloseNotification}
+																variant="outlined"
+																color="error"
+																sx={{
+																	textTransform: "capitalize",
+																	borderRadius: "8px",
+																	fontWeight: "500",
+																	fontSize: "13px",
+																	padding: "11px 30px",
+																}}
+															>
+																Cancel
+															</Button>
 
-                                                            <Button
-                                                                type="submit"
-                                                                variant="contained"
-                                                                sx={{
-                                                                    textTransform: "capitalize",
-                                                                    borderRadius: "8px",
-                                                                    fontWeight: "500",
-                                                                    fontSize: "13px",
-                                                                    padding: "11px 30px",
-                                                                    color: "#fff !important",
-                                                                }}
-                                                            >
-                                                                Move
-                                                            </Button>
-                                                        </Box>
-                                                    </Grid>
-                                                </Grid>
-                                            </Box>
-                                        </Box>
-                                    </Box>
-                                </Box>
-                            </BootstrapDialog>
-                        </Card> */}
+															<Button
+																type="submit"
+																variant="contained"
+																sx={{
+																	textTransform: "capitalize",
+																	borderRadius: "8px",
+																	fontWeight: "500",
+																	fontSize: "13px",
+																	padding: "11px 30px",
+																	color: "#fff !important",
+																}}
+															>
+																Move
+															</Button>
+														</Box>
+													</Grid>
+												</Grid>
+											</Box>
+										</Box>
+									</Box>
+								</Box>
+							</BootstrapDialog>
+						</Card> */}
 						<Box>
 							<DndContext
 								modifiers={[restrictToVerticalAxis]}
@@ -1050,6 +1008,7 @@ export default function Page({ projectName }: { projectName: string }) {
 											padding: { xs: "8px", sm: "10px", lg: "15px" },
 										}}
 										className="rmui-card"
+										key={sprint._id}
 									>
 										<Box
 											display="flex"
@@ -1100,14 +1059,14 @@ export default function Page({ projectName }: { projectName: string }) {
 										<Accordion
 											expanded={
 												Array.isArray(expanded)
-													? expanded.includes(sprint)
-													: expanded === sprint
+													? expanded.includes(sprint.name)
+													: expanded === sprint.name
 											}
-											onChange={handleAccordionChange(sprint)}
+											onChange={handleAccordionChange(sprint.name)}
 											className="accordionItem"
 											sx={{
 												backgroundColor:
-													expanded === sprint ? "#e9ebee" : "inherit",
+													expanded === sprint.name ? "#e9ebee" : "inherit",
 												"&:hover": {
 													backgroundColor: "#e9ebee",
 												},
@@ -1123,7 +1082,7 @@ export default function Page({ projectName }: { projectName: string }) {
 												id="panel4-header"
 												sx={{ fontWeight: "500", fontSize: "15px" }}
 											>
-												{sprint}
+												{sprint.name}
 											</AccordionSummary>
 											<AccordionDetails>
 												<Stack spacing={1}>
@@ -1142,151 +1101,7 @@ export default function Page({ projectName }: { projectName: string }) {
 															}
 														>
 															<Stack spacing={1}>
-																{newbacklogs.map((newbacklog, index) => (
-																	<>
-																		<Item
-																			className="backlogItem"
-																			style={{
-																				padding: "0px 0px 0px 0px",
-																			}}
-																		>
-																			<Table
-																				sx={{
-																					borderBottom: "none !important",
-																				}}
-																			>
-																				<TableBody>
-																					<TableRow>
-																						<TableCell
-																							style={{
-																								border: "none",
-																								alignItems: "center",
-																								justifyContent: "center",
-																							}}
-																						>
-																							<div
-																								style={{
-																									paddingTop: "5px",
-																									display: "flex",
-																									justifyContent: "center",
-																								}}
-																							>
-																								<svg
-																									width="20px"
-																									height="20px"
-																									style={{
-																										marginRight: "5px",
-																									}}
-																									viewBox="0 0 16 16"
-																									version="1.1"
-																									xmlns="http://www.w3.org/2000/svg"
-																								>
-																									<defs></defs>
-																									<g
-																										id="Page-1"
-																										stroke="none"
-																										strokeWidth="1"
-																										fill="none"
-																										fillRule="evenodd"
-																									>
-																										<g id="task">
-																											<g
-																												id="Task"
-																												transform="translate(1.000000, 1.000000)"
-																											>
-																												<rect
-																													id="Rectangle-36"
-																													fill="#4BADE8"
-																													x="0"
-																													y="0"
-																													width="14"
-																													height="14"
-																													rx="2"
-																												></rect>
-																												<g
-																													id="Page-1"
-																													transform="translate(4.000000, 4.500000)"
-																													stroke="#FFFFFF"
-																													strokeWidth="2"
-																													strokeLinecap="round"
-																												>
-																													<path
-																														d="M2,5 L6,0"
-																														id="Stroke-1"
-																													></path>
-																													<path
-																														d="M2,5 L0,3"
-																														id="Stroke-3"
-																													></path>
-																												</g>
-																											</g>
-																										</g>
-																									</g>
-																								</svg>
-																							</div>
-																						</TableCell>
-																						<TableCell
-																							style={{
-																								border: "none",
-																							}}
-																						>
-																							{loading ? (
-																								<LinearProgress
-																									sx={{
-																										width: "100vh",
-																										color: "white",
-																									}}
-																									color="secondary"
-																								/>
-																							) : (
-																								<Input
-																									id="issueName"
-																									placeholder="Backlog name.."
-																									sx={{
-																										width: "100%",
-																										color: "white",
-																									}}
-																									aria-label="Name"
-																									value={issueName}
-																									onChange={(event) =>
-																										setIssueName(
-																											event.target.value
-																										)
-																									}
-																									onKeyDown={(event) => {
-																										if (event.key === "Enter") {
-																											handleBacklogSubmit();
-																										}
-																									}}
-																									autoFocus
-																								/>
-																							)}
-																						</TableCell>
-																						<TableCell
-																							style={{
-																								border: "none",
-																								display: "flex",
-																								flexDirection: "row",
-																								alignItems: "center",
-																							}}
-																						></TableCell>
-																						<TableCell
-																							style={{
-																								border: "none",
-																							}}
-																						></TableCell>
-																					</TableRow>
-																				</TableBody>
-																			</Table>
-																		</Item>
-																	</>
-																))}
-																<Button
-																	className="createIssueBtn"
-																	onClick={handleCreateBacklog}
-																>
-																	+ Create Issue
-																</Button>
+																<CreateBacklog />
 															</Stack>
 														</DndContext>
 													</AccordionDetails>
@@ -1370,6 +1185,9 @@ export default function Page({ projectName }: { projectName: string }) {
 																		</Button>
 
 																		<Button
+																			onClick={() =>
+																				handleDeleteSprint(sprint._id)
+																			}
 																			type="submit"
 																			variant="contained"
 																			sx={{
@@ -1394,14 +1212,23 @@ export default function Page({ projectName }: { projectName: string }) {
 									</Card>
 								))}
 								<Box>
-									<Button
-										variant="outlined"
-										size="medium"
-										sx={{ marginBottom: "10px" }}
-										onClick={handleCreateSprint}
-									>
-										Create Sprint
-									</Button>
+									{loading ? (
+										<LinearProgress
+											sx={{
+												width: "100%",
+											}}
+											color="success"
+										/>
+									) : (
+										<Button
+											variant="outlined"
+											size="medium"
+											sx={{ marginBottom: "10px" }}
+											onClick={handleCreateSprint}
+										>
+											Create Sprint
+										</Button>
+									)}
 								</Box>
 								<Card
 									sx={{
@@ -1456,7 +1283,189 @@ export default function Page({ projectName }: { projectName: string }) {
 													Backlog
 												</AccordionSummary>
 												<AccordionDetails>
-													<BacklogList backlogs={backlogs}></BacklogList>
+													{/* <BacklogList backlogs={backlogs}></BacklogList> */}
+													<Stack spacing={1}>
+														<AccordionDetails>
+															<DndContext
+																modifiers={[restrictToVerticalAxis]}
+																sensors={sensors}
+																collisionDetection={closestCorners}
+																onDragStart={handleOnDragStart}
+																onDragEnd={(event) =>
+																	handleDragEnd(
+																		event as {
+																			active: { id: string };
+																			over: { id: string };
+																		}
+																	)
+																}
+															>
+																<BacklogList backlogs={backlogs}></BacklogList>
+																<DragOverlay dropAnimation={dropAnimation}>
+																	{!activeDragItemData && null}
+																	{
+																		<div
+																			className="backlogItem"
+																			style={{ padding: "0px 0px 0px 0px" }}
+																		>
+																			<Table
+																				sx={{
+																					borderBottom: "none !important",
+																				}}
+																			>
+																				<TableBody>
+																					<TableRow>
+																						<TableCell
+																							style={{
+																								border: "none",
+																								alignItems: "center",
+																								justifyContent: "center",
+																							}}
+																						>
+																							<div
+																								style={{
+																									paddingTop: "5px",
+																									display: "flex",
+																									justifyContent: "center",
+																								}}
+																							>
+																								<svg
+																									width="20px"
+																									height="20px"
+																									style={{
+																										marginRight: "5px",
+																									}}
+																									viewBox="0 0 16 16"
+																									version="1.1"
+																									xmlns="http://www.w3.org/2000/svg"
+																								>
+																									<defs></defs>
+																									<g
+																										id="Page-1"
+																										stroke="none"
+																										strokeWidth="1"
+																										fill="none"
+																										fillRule="evenodd"
+																									>
+																										<g id="task">
+																											<g
+																												id="Task"
+																												transform="translate(1.000000, 1.000000)"
+																											>
+																												<rect
+																													id="Rectangle-36"
+																													fill="#4BADE8"
+																													x="0"
+																													y="0"
+																													width="14"
+																													height="14"
+																													rx="2"
+																												></rect>
+																												<g
+																													id="Page-1"
+																													transform="translate(4.000000, 4.500000)"
+																													stroke="#FFFFFF"
+																													strokeWidth="2"
+																													strokeLinecap="round"
+																												>
+																													<path
+																														d="M2,5 L6,0"
+																														id="Stroke-1"
+																													></path>
+																													<path
+																														d="M2,5 L0,3"
+																														id="Stroke-3"
+																													></path>
+																												</g>
+																											</g>
+																										</g>
+																									</g>
+																								</svg>
+																							</div>
+																						</TableCell>
+																						<TableCell
+																							style={{
+																								border: "none",
+																							}}
+																							sx={{ width: "50%" }}
+																						>
+																							<Link
+																								className="hover-underlined"
+																								color="inherit"
+																								href=""
+																							></Link>
+																						</TableCell>
+																						<TableCell
+																							style={{
+																								border: "none",
+																								display: "flex",
+																								flexDirection: "row",
+																							}}
+																						>
+																							<Select
+																								labelId="product-type-label"
+																								id="product-type"
+																								className="epicSelectBg"
+																								size="small"
+																								style={{
+																									marginRight: "5px",
+																								}}
+																								sx={{
+																									"& fieldset": {
+																										maxWidth: "120px",
+																									},
+																									"& .MuiSelect-select": {
+																										overflow: "hidden",
+																										textOverflow: "ellipsis",
+																										whiteSpace: "nowrap",
+																									},
+																								}}
+																							>
+																								<MenuItem value={0}>
+																									epic
+																								</MenuItem>
+																								<MenuItem value={1}>1</MenuItem>
+																								<MenuItem value={2}>2</MenuItem>
+																								<MenuItem value={3}>3</MenuItem>
+																								<MenuItem value={4}>4</MenuItem>
+																							</Select>
+																							<Select
+																								labelId="product-type-label"
+																								className="progressSelectBg"
+																								id="product-type"
+																								size="small"
+																								sx={{
+																									"& fieldset": {},
+																									"& .MuiSelect-select": {
+																										overflow: "hidden",
+																										textOverflow: "ellipsis",
+																										whiteSpace: "nowrap",
+																									},
+																								}}
+																							>
+																								<MenuItem value={0}>
+																									progress
+																								</MenuItem>
+																								<MenuItem value={1}>1</MenuItem>
+																								<MenuItem value={2}>2</MenuItem>
+																								<MenuItem value={3}>3</MenuItem>
+																								<MenuItem value={4}>4</MenuItem>
+																							</Select>
+																						</TableCell>
+																						<TableCell
+																							style={{
+																								border: "none",
+																							}}
+																						></TableCell>
+																					</TableRow>
+																				</TableBody>
+																			</Table>
+																		</div>
+																	}
+																</DragOverlay>
+															</DndContext>
+														</AccordionDetails>
+													</Stack>
 												</AccordionDetails>
 											</Accordion>
 										</Typography>
